@@ -5,10 +5,11 @@
   "homepage": "Tutorials/MapRendering/ptv-vectormap-react-app",
   "dependencies": {
     
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-map-gl": "^5.3.7",
-    "react-scripts": "4.0.3"
+    "mapbox-gl": "^1.13.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-map-gl": "^7.0.25",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
