@@ -1,11 +1,13 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import "./index.css";
 import App from "./App";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(
+  document.getElementById('root')
+);
+root.render(
   <React.StrictMode>
     <App/>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
