@@ -1,7 +1,7 @@
 
 import { getMapStyle, initialMapStyle } from "./getMapStyle";
 import { useEffect, useState, useCallback } from "react";
-import ReactMapGL, { NavigationControl } from "react-map-gl";
+import Map, { NavigationControl } from "react-map-gl";
 
 // use the standard map style provided by PTV
 const MAP_STYLE_URL = "https://vectormaps-resources.myptv.com/styles/latest/standard.json";
@@ -11,10 +11,19 @@ const navControlStyle = {
   top: 10
 };
 
+// Initializes to Karlsruhe
+const initViewport = {
+  longitude: 8.4055677,
+  latitude: 49.0070036,
+  zoom: 10,
+  pitch: 0,
+  bearing: 0,
+};
+
+
 export const VectorMap = (props) => {
 
   const [mapStyle, setMapStyle] = useState(initialMapStyle);
-  const [viewport, setViewport] = useState(props.viewport);
 
   useEffect(() => {
     getMapStyle(MAP_STYLE_URL).then(setMapStyle);
@@ -33,14 +42,15 @@ export const VectorMap = (props) => {
     }
   );
 
-  return <ReactMapGL
+
+
+  return <Map
     height="100%"
     width="100%"
     mapStyle={mapStyle}
-    {...viewport}
-    onViewportChange={setViewport}
+    initialViewState={initViewport}
     transformRequest={(url, resourceType) => getTransformRequest(url, resourceType)}
   >
     <NavigationControl style={navControlStyle}/>
-  </ReactMapGL>;
+  </Map>;
 };
